src/lib/analysers/TorchHub/requirements.txt
# base ----------------------------------------
Cython
matplotlib>=3.2.2
# numpy>=1.18.5
opencv-python>=4.1.2
Pillow
# PyYAML>=5.3.1
# scipy>=1.4.1
# tensorboard>=2.2
requests
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0

# logging -------------------------------------
# wandb

# plotting ------------------------------------
seaborn>=0.11.0
pandas

# extras --------------------------------------
# thop  # FLOPS computation
# pycocotools>=2.0  # COCO mAP
